updated the clumping function to use 1/z

+++ clump.py
@@ -5,6 +5,7 @@ import argparse
 import math
 import os
 
+
 parser = argparse.ArgumentParser(description="Extract and clump top hits")
 parser.add_argument("--bcf", required=True)
 parser.add_argument("--bcftools_binary", default="bcftools")
@@ -14,15 +15,20 @@ parser.add_argument("--clump_pval", type=float, default=5e-8)
 parser.add_argument("--clump_kb", type=int, default=10000)
 parser.add_argument("--clump_r2", type=float, default=0.001)
 parser.add_argument("--out", required=True)
+parser.add_argument("--debug", required=False, default=False)
 args = parser.parse_args()
 
+
+# Extract the SNPs that are significant at the clump_pval
+# Some of these might be zero - would really like the one with the biggest test statistic
+# Use 1/zscore instead of pval - this will choose the best test statistic
 cmd = " ".join(
     [
         "{0} view -i 'FORMAT/LP>{1}' {2} |",
-        "{0} query -f'%ID[ %LP]\n' |",
-        'awk \'BEGIN {{print "SNP P"}};',
-        "{{print $1, 10^-$2}}' |",
-        "awk '!seen[$1]++' > {3}.tophits",
+        "{0} query -f'%ID[ %ES %SE %LP]\n' |",
+        'awk \' function abs(v) {{return v < 0 ? -v : v}} BEGIN {{print "SNP P"}};',
+        "{{print $1, $3/abs($2)}}' |",
+        "awk '!seen[$1]++' > {3}.tophits"
     ]
 ).format(
     args.bcftools_binary, -math.log10(args.clump_pval), args.bcf, args.out
@@ -41,7 +47,7 @@ cmd = " ".join(
     args.out + ".tophits",
     args.clump_kb,
     args.clump_r2,
-    args.clump_pval,
+    1,
 )
 os.system(cmd)
 
@@ -62,8 +68,9 @@ else:
 
 os.system("touch " + args.out)
 
-[
-    os.remove(args.out + x)
-    for x in [".tophits.clumped", ".tophits.log", ".tophits.nosex", ".tophits"]
-    if os.path.isfile(args.out + x)
-]
+if args.debug is False:
+    [
+        os.remove(args.out + x)
+        for x in [".tophits.clumped", ".tophits.log", ".tophits.nosex", ".tophits"]
+        if os.path.isfile(args.out + x)
+    ]
